Broke the app up into some higher-order functions in a top-level Trampoline module and then core "inner" app logic.
Added the basics of a version that uses rewriting and step counting rather than gas, but I haven't used it yet. The two are inter-translatable. Gas is arguably easier to understand (and write, if you can stomach monads by hand in Elm), but rewriting offers intermediate states. Leaning towards making the latter the "core" idea.